Commands executed via system() now redirect via IPC::Run, and log via printMsg()
We can use IPC::Run to spawn external commands and to gather the output. This
allows consistent handling of STDOUT and STDERR results from the programs.

Also, if($Debug) we now log the command output.

This patch intends to make absolutely no changes in behavior, but to pave the
way for beneficial future changes.

This patch does not touch commands executed with ``, but it probably should.